+++ sklearn/neighbors/tests/test_neighbors.py
@@ -1,9 +1,11 @@
 from itertools import product
+import pickle
 
 import numpy as np
 from scipy.sparse import (bsr_matrix, coo_matrix, csc_matrix, csr_matrix,
                           dok_matrix, lil_matrix)
 
+from sklearn import metrics
 from sklearn.cross_validation import train_test_split
 from sklearn.utils.testing import assert_array_almost_equal
 from sklearn.utils.testing import assert_array_equal
@@ -98,6 +100,32 @@ def test_unsupervised_inputs():
 
         assert_array_almost_equal(dist1, dist2)
         assert_array_almost_equal(ind1, ind2)
+
+
+def test_unsupervisd_knn_distance():
+    """Tests unsupervised NearestNeighbors with a distance matrix."""
+    X = rng.random_sample((3, 4))  # Must not be square for tests below.
+    D = metrics.pairwise_distances(X, metric='euclidean')
+    # As a feature matrix (n_samples by n_features)
+    nbrs_X = neighbors.NearestNeighbors(n_neighbors=3)
+    nbrs_X.fit(X)
+    dist_X, ind_X = nbrs_X.kneighbors(X)
+    # As a dense distance matrix (n_samples by n_samples)
+    nbrs_D = neighbors.NearestNeighbors(n_neighbors=3, algorithm='brute',
+                                        metric='precomputed')
+    nbrs_D.fit(D)
+    dist_D, ind_D = nbrs_D.kneighbors(D)
+    # Assert that they give the same neighbors
+    assert_array_almost_equal(dist_X, dist_D)
+    assert_array_almost_equal(ind_X, ind_D)
+    # Test pickling
+    pickled_classifier = pickle.dumps(nbrs_D)
+    unpickled_classifier = pickle.loads(pickled_classifier)
+    unpickled_dist_D, unpickled_ind_D = unpickled_classifier.kneighbors(D)
+    assert_array_almost_equal(unpickled_dist_D, dist_D)
+    assert_array_almost_equal(unpickled_ind_D, ind_D)
+    # Must raise a ValueError if the matrix is not square
+    assert_raises(ValueError, nbrs_D.kneighbors, X)  # X is not square
 
 
 def test_unsupervised_radius_neighbors(n_samples=20, n_features=5,
@@ -942,8 +970,7 @@ def test_non_euclidean_kneighbors():
         nbrs_graph = neighbors.radius_neighbors_graph(
             X, radius, metric=metric).toarray()
         nbrs1 = neighbors.NearestNeighbors(metric=metric, radius=radius).fit(X)
-        assert_array_equal(nbrs_graph,
-                           nbrs1.radius_neighbors_graph(X).toarray())
+        assert_array_equal(nbrs_graph, nbrs1.radius_neighbors_graph(X).A)
 
     # Raise error when wrong parameters are supplied,
     X_nbrs = neighbors.NearestNeighbors(3, metric='manhattan')